+++ API/Controllers/LecturesController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using AuthenticationLibrary.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 using Model.Repositories;
 using Services;
@@ -10,39 +11,78 @@ namespace API.Controllers
     [Route("api/Lectures")]
     public class LecturesController : Controller
     {
-        public readonly IService service;
+        private readonly IService service;
+        private readonly IAuthService authService;
 
-        public LecturesController(IService service)
+        public LecturesController(IService service, IAuthService authService)
         {
             this.service = service;
+            this.authService = authService;
         }
 
         // GET: api/Lectures
         [HttpGet]
-        public async Task<IEnumerable<Lecture>> Get()
+        public async Task<IActionResult> Get()
         {
-            return await service.GetLectures();
+            var admin = await authService.ValidateAdmin(Request);
+            var teacher = await authService.ValidateTeacher(Request);
+
+            if (admin != null || teacher != null)
+            {
+                var lectures = await service.GetLectures();
+                return Ok(lectures);
+            }
+
+            return Unauthorized();
         }
 
         // POST: api/Lectures
         [HttpPost]
-        public async Task<int> Post([FromBody]Lecture lecture)
+        public async Task<IActionResult> Post([FromBody]Lecture lecture)
         {
-            return await service.AddLecture(lecture);
+
+            var admin = await authService.ValidateAdmin(Request);
+            var teacher = await authService.ValidateTeacher(Request);
+
+            if (admin != null || teacher != null)
+            {
+                var lectureID = await service.AddLecture(lecture);
+                return Ok(lectureID);
+            }
+
+            return Unauthorized();
         }
 
         // PUT: api/Lectures/5
         [HttpPut("{id}")]
-        public async Task<bool> Put(int id, [FromBody]Lecture lecture)
+        public async Task<IActionResult> Put(int id, [FromBody]Lecture lecture)
         {
-            return await service.UpdateLecture(lecture, id);
+            var admin = await authService.ValidateAdmin(Request);
+            var teacher = await authService.ValidateTeacher(Request);
+
+            if (admin != null || teacher != null)
+            {
+                var success = await service.UpdateLecture(lecture, id);
+                return Ok(success);
+            }
+
+            return Unauthorized();
         }
 
         // DELETE: api/ApiWithActions/5
         [HttpDelete("{id}")]
-        public async Task<bool> Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
-            return await service.DeleteLecture(id);
+            var admin = await authService.ValidateAdmin(Request);
+            var teacher = await authService.ValidateTeacher(Request);
+
+            if (admin != null || teacher != null)
+            {
+                var success = await service.DeleteLecture(id);
+                return Ok(success);
+            }
+
+            return Unauthorized();
         }
     }
 }
