+++ API/Controllers/StudentsController.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using AuthenticationLibrary.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 using Model.Repositories;
 using Services;
@@ -13,39 +14,77 @@ namespace API.Controllers
     {
         private readonly IService service;
         private readonly IUsersFacade usersFacade;
+        private readonly IAuthService authService;
 
-        public StudentsController(IService service, IUsersFacade usersFacade)
+        public StudentsController(IService service, IUsersFacade usersFacade, IAuthService authService)
         {
             this.service = service;
             this.usersFacade = usersFacade;
+            this.authService = authService;
         }
 
         // GET: api/Students
         [HttpGet]
-        public async Task<IEnumerable<Student>> Get()
+        public async Task<IActionResult> Get()
         {
-            return await service.GetStudents();
+            var admin = await authService.ValidateAdmin(Request);
+            var teacher = await authService.ValidateTeacher(Request);
+
+            if (admin != null || teacher != null)
+            {
+                var students = await service.GetStudents();
+                return Ok(students);
+            }
+
+            return Unauthorized();
         }
 
         // POST: api/Students
         [HttpPost]
-        public async Task<int> Post([FromBody]Student student)
+        public async Task<IActionResult> Post([FromBody]Student student)
         {
-            return await usersFacade.AddStudentUser(student);
+            var admin = await authService.ValidateAdmin(Request);
+            var teacher = await authService.ValidateTeacher(Request);
+
+            if (admin != null || teacher != null)
+            {
+                var studentID = await usersFacade.AddStudentUser(student);
+                return Ok(studentID);
+            }
+
+            return Unauthorized();
         }
 
         // PUT: api/Students/5
         [HttpPut("{id}")]
-        public async Task<bool> Put(int id, [FromBody]Student student)
+        public async Task<IActionResult> Put(int id, [FromBody]Student student)
         {
-            return await service.UpdateStudent(student, id);
+            var admin = await authService.ValidateAdmin(Request);
+            var teacher = await authService.ValidateTeacher(Request);
+
+            if (admin != null || teacher != null)
+            {
+                var success = await service.UpdateStudent(student, id);
+                return Ok(success);
+            }
+
+            return Unauthorized();
         }
 
         // DELETE: api/ApiWithActions/5
         [HttpDelete("{id}")]
-        public async Task<bool> Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
-            return await service.DeleteStudent(id);
+            var admin = await authService.ValidateAdmin(Request);
+            var teacher = await authService.ValidateTeacher(Request);
+
+            if (admin != null || teacher != null)
+            {
+                var success = await service.DeleteStudent(id);
+                return Ok(success);
+            }
+
+            return Unauthorized();
         }
     }
 }
